Split all API endpoints into separate files
