+++ etna/api/urls/images.py
@@ -1,0 +1,19 @@
+from wagtail.images.api.v2.views import ImagesAPIViewSet
+
+
+class CustomImagesAPIViewSet(ImagesAPIViewSet):
+    body_fields = ImagesAPIViewSet.body_fields + [
+        "title",
+        "file",
+        "copyright",
+        "is_sensitive",
+        "custom_sensitive_image_warning",
+        "tags",
+        "transcription_heading",
+        "transcription",
+        "translation_heading",
+        "translation",
+        "record",
+        "record_dates",
+        "description",
+    ]
