+++ etna/api/urls/pages.py
@@ -1,29 +1,19 @@
 import logging
 
 from django.conf import settings
-from django.contrib.contenttypes.models import ContentType
 from django.db.models import Q
 from django.http import Http404
 from django.shortcuts import redirect
-from django.urls import path
 from django.utils.crypto import constant_time_compare
 from rest_framework import status
 from rest_framework.response import Response
-from wagtail.api.v2.router import WagtailAPIRouter
 from wagtail.api.v2.utils import BadRequestError, get_object_detail_url
-from wagtail.api.v2.views import BaseAPIViewSet, PagesAPIViewSet
+from wagtail.api.v2.views import PagesAPIViewSet
 from wagtail.contrib.redirects.models import Redirect
-from wagtail.images.api.v2.views import ImagesAPIViewSet
 from wagtail.models import Page, PageViewRestriction, Site
 from wagtail.search.backends.database.postgres.postgres import PostgresSearchResults
-from wagtail_headless_preview.models import PagePreview
-from wagtailmedia.api.views import MediaAPIViewSet
 
-from etna.blog.models import BlogIndexPage, BlogPage, BlogPostPage
 from etna.core.serializers.pages import DefaultPageSerializer
-from etna.core.serializers.redirects import RedirectSerializer
-
-from .filters import AuthorFilter, PublishedDateFilter
 
 logger = logging.getLogger(__name__)
 
@@ -242,233 +232,3 @@ class CustomPagesAPIViewSet(PagesAPIViewSet):
                 return page
 
         return super().find_object(queryset, request)
-
-
-class PagePreviewAPIViewSet(PagesAPIViewSet):
-    known_query_parameters = PagesAPIViewSet.known_query_parameters.union(
-        ["content_type", "token"]
-    )
-
-    def listing_view(self, request):
-        page = self.get_object()
-        serializer = self.get_serializer(page)
-        return Response(serializer.data)
-
-    def detail_view(self, request, pk):
-        page = self.get_object()
-        serializer = self.get_serializer(page)
-        return Response(serializer.data)
-
-    def get_object(self):
-        if "content_type" not in self.request.GET:
-            raise BadRequestError("content_type not specified")
-        if "token" not in self.request.GET:
-            raise BadRequestError("token not specified")
-
-        app_label, model = self.request.GET["content_type"].split(".")
-        content_type = ContentType.objects.get(app_label=app_label, model=model)
-
-        try:
-            page_preview = PagePreview.objects.get(
-                content_type=content_type, token=self.request.GET["token"]
-            )
-        except PagePreview.DoesNotExist:
-            raise BadRequestError("Page preview does not exist")
-        page = page_preview.as_page()
-        if not page.pk:
-            # fake primary key to stop API URL routing from complaining
-            page.pk = 0
-
-        return page
-
-
-class CustomImagesAPIViewSet(ImagesAPIViewSet):
-    body_fields = ImagesAPIViewSet.body_fields + [
-        "title",
-        "file",
-        "copyright",
-        "is_sensitive",
-        "custom_sensitive_image_warning",
-        "tags",
-        "transcription_heading",
-        "transcription",
-        "translation_heading",
-        "translation",
-        "record",
-        "record_dates",
-        "description",
-    ]
-
-
-class BlogsAPIViewSet(CustomPagesAPIViewSet):
-    model = BlogPage
-
-    def top_level_blogs_list_view(self, request):
-        queryset = self.get_queryset()
-        restricted_pages = [
-            restriction.page
-            for restriction in PageViewRestriction.objects.all().select_related("page")
-            if not restriction.accept_request(self.request)
-        ]
-        for restricted_page in restricted_pages:
-            queryset = queryset.not_descendant_of(restricted_page, inclusive=True)
-        blog_post_counts = {}
-        for blog in queryset:
-            # Ignore all "sub-blogs" (BlogPages which are children of other BlogPages)
-            queryset = queryset.not_descendant_of(blog, inclusive=False)
-            blog_posts = BlogPostPage.objects.all().live().descendant_of(blog).count()
-            blog_post_counts[blog.id] = blog_posts
-        serializer = DefaultPageSerializer(
-            queryset, required_api_fields=["custom_type_label"], many=True
-        )
-        blogs = sorted(serializer.data, key=lambda x: x["title"])
-        blogs = [blog | {"posts": blog_post_counts[blog["id"]]} for blog in blogs]
-        top_level_queryset = BlogIndexPage.objects.all().live()
-        top_level = DefaultPageSerializer(top_level_queryset, many=True)
-        blogs = top_level.data + blogs
-        return Response(blogs)
-
-    def blog_index_view(self, request):
-        queryset = BlogIndexPage.objects.all().live()
-        blog_index = DefaultPageSerializer(queryset, many=True)
-        return Response(blog_index.data[0] if len(blog_index.data) else None)
-
-    @classmethod
-    def get_urlpatterns(cls):
-        """
-        This returns a list of URL patterns for the endpoint
-        """
-        return [
-            path("", cls.as_view({"get": "listing_view"}), name="blogs_list"),
-            path(
-                "index/",
-                cls.as_view({"get": "blog_index_view"}),
-                name="blogs_index",
-            ),
-            path(
-                "top/",
-                cls.as_view({"get": "top_level_blogs_list_view"}),
-                name="top_level_blogs_list",
-            ),
-        ]
-
-
-class BlogPostsAPIViewSet(CustomPagesAPIViewSet):
-    filter_backends = [
-        PublishedDateFilter,
-        AuthorFilter,
-    ] + CustomPagesAPIViewSet.filter_backends  # Needs to be last as it includes SearchFilter which needs to be last
-    known_query_parameters = CustomPagesAPIViewSet.known_query_parameters.union(
-        [
-            "year",
-            "month",
-            "day",
-            "author",
-        ]
-    )
-    model = BlogPostPage
-
-    def count_view(self, request):
-        queryset = self.get_queryset()
-        self.check_query_parameters(queryset)
-        queryset = self.filter_queryset(queryset).public()
-        years = set(queryset.values_list("published_date__year", flat=True))
-        years_count = [
-            {
-                "year": year,
-                "months": [
-                    {
-                        "month": month,
-                        "posts": queryset.filter(
-                            **{
-                                "published_date__year": year,
-                                "published_date__month": month,
-                            }
-                        ).count(),
-                    }
-                    for month in sorted(
-                        set(
-                            queryset.filter(
-                                **{"published_date__year": year}
-                            ).values_list("published_date__month", flat=True)
-                        )
-                    )
-                ],
-                "posts": queryset.filter(**{"published_date__year": year}).count(),
-            }
-            for year in sorted(years)
-        ]
-        return Response(years_count)
-
-    def author_view(self, request):
-        queryset = self.get_queryset()
-        self.check_query_parameters(queryset)
-        queryset = self.filter_queryset(queryset)
-        authors = set(queryset.values_list("author_tags__author"))
-        serializer = DefaultPageSerializer()
-        authors_count = []
-        for author in authors:
-            if author[0] is not None:
-                author_item = (
-                    queryset.filter(author_tags__author=author)
-                    .first()
-                    .author_tags.filter(author=author)
-                    .first()
-                    .author
-                )
-                authors_count.append(
-                    {
-                        "author": serializer.to_representation(author_item),
-                        "posts": queryset.filter(author_tags__author=author).count(),
-                    }
-                )
-        return Response(sorted(authors_count, key=lambda x: x["posts"], reverse=True))
-
-    @classmethod
-    def get_urlpatterns(cls):
-        """
-        This returns a list of URL patterns for the endpoint
-        """
-        return [
-            path("", cls.as_view({"get": "listing_view"}), name="listing"),
-            path("count/", cls.as_view({"get": "count_view"}), name="count"),
-            path("authors/", cls.as_view({"get": "author_view"}), name="authors"),
-        ]
-
-
-class RedirectsAPIViewSet(BaseAPIViewSet):
-    model = Redirect
-
-    def listing_view(self, request):
-        queryset = self.get_queryset()
-        self.check_query_parameters(queryset)
-        queryset = self.filter_queryset(queryset)
-        queryset = self.paginate_queryset(queryset)
-        serializer = RedirectSerializer(queryset, many=True)
-        return self.get_paginated_response(serializer.data)
-
-    def detail_view(self, request, pk):
-        instance = self.get_object()
-        serializer = RedirectSerializer(instance)
-        return Response(serializer.data)
-
-    def find_object(self, queryset, request):
-        if "path" in request.GET:
-            # Taken from https://github.com/wagtail/wagtail/blob/main/wagtail/contrib/redirects/models.py#L165-L169
-            path = request.GET["path"]
-            path = path.strip()
-            if not path.startswith("/"):
-                path = "/" + path
-            if path.endswith("/") and len(path) > 1:
-                path = path[:-1]
-            return queryset.get(old_path=path)
-
-
-api_router = WagtailAPIRouter("wagtailapi")
-api_router.register_endpoint("pages", CustomPagesAPIViewSet)
-api_router.register_endpoint("page_preview", PagePreviewAPIViewSet)
-api_router.register_endpoint("images", CustomImagesAPIViewSet)
-api_router.register_endpoint("media", MediaAPIViewSet)
-api_router.register_endpoint("blogs", BlogsAPIViewSet)
-api_router.register_endpoint("blog_posts", BlogPostsAPIViewSet)
-api_router.register_endpoint("redirects", RedirectsAPIViewSet)
